feat: secure journal endpoints with Basic Auth

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -29,7 +29,7 @@ public class JournalEntryService {
            journalEntry.setDate(LocalDateTime.now());
            JournalEntry saved = journalEntryRepository.save(journalEntry);
            user.getJournalEntries().add(saved);
-           userService.saveEntry(user);
+           userService.saveUser(user);
        }catch (Exception e)
        {
            System.out.println(e);
@@ -50,10 +50,25 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x-> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+   public boolean deleteById(ObjectId id, String userName){
+        boolean removed=false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed= user.getJournalEntries().removeIf(x-> x.getId().equals(id));
+            if (removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }catch (Exception e)
+        {
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting entry",e);
+        }
+        return removed;
     }
+
+//    public List<JournalEntry> findByUserName(String userName){
+//
+//    }
 }
